+++ CHANGELOG.md
@@ -5,6 +5,15 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/)
 and adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.6.0] - 2025-12-18
+
+### Added
+- `HightideConfigContext` for storing Look and Feel Variables
+- `HightideProvider` to bundle all hightide contexts in one provider
+
+### Changed
+- `Tables` to no longer round their column size
+
 ## [0.5.4] - 2025-12-18
 
 ### Fixed